Changed to match recent soy job & opengl changes
Former-commit-id: b37c368e37d33e5142d92b157d2cc178a3968949

+++ src/TFilterStageGatherRects.cpp
@@ -242,7 +242,7 @@ void TFilterStage_MakeRectAtlas::CreateBlitResources()
 		Mesh.mVertexes[1].uv = vec2f( 1, 0);
 		Mesh.mVertexes[2].uv = vec2f( 1, 1);
 		Mesh.mVertexes[3].uv = vec2f( 0, 1);
-		Array<GLshort> Indexes;
+		Array<size_t> Indexes;
 		Indexes.PushBack( 0 );
 		Indexes.PushBack( 1 );
 		Indexes.PushBack( 2 );
